Method Updates
-Check balance method added to the customer class
-Add Customer error checking added to Flow class
-Updated the create credit function to check for the credit balance and check if the credit number already exists in both service provider

+++ Java/OOPproject/teleCompanyPKG/ServiceProvider.java
@@ -56,7 +56,7 @@ public abstract class ServiceProvider {
 	}
 	
 	//Phone Credit Creation Method, will be overrided in child classes
-	public boolean createPhoneCredit(int cardNum, int balance) {
+	public boolean createPhoneCredit(int cardNum, float balance) {
 		return true;
 	};
